+++ managers/entities/user/User.manager.js
@@ -1,3 +1,5 @@
+const { getSuperAdmins } = require('../../../exportSuperAdmins');
+const superAdmins = getSuperAdmins();
 module.exports = class User { 
 
     constructor({utils, cache, config, cortex, managers, validators, mongomodels }={}){
@@ -7,16 +9,13 @@ module.exports = class User {
         this.mongomodels         = mongomodels;
         this.tokenManager        = managers.token;
         this.usersCollection     = "users";
-        this.httpExposed         = ['post=createUser', 'get=getUser', 'put=updateUser','delete=deleteUser', 'get=listUsers'];
+        this.httpExposed         = ['post=createUser', 'get=getUser', 'put=updateUser','delete=deleteUser', 'get=listUsers', 'post=loginUser'];
         this.utils= utils;
     }
 
     async createUser({username, email, password}){  
         const user = {username, email, password};
 
-        // Data validation
-        let result = await this.validators.user.createUser(user);
-        if(result) return result;
         
         try {
         // Check if user already exists
@@ -26,10 +25,16 @@ module.exports = class User {
         }
          // Hash the password before saving
         user.password = await this.utils.hashPassword(password);
+
             // Save the user to the database
-        let createdUser = await this.mongomodels.UserModel.create(user);
+        let createdUser;
+        try{
+         createdUser = await this.mongomodels.UserModel.create(user);}
+        catch(err){
+            return this.utils.formatMongooseError(err);
+        }
         // Creation Logic
-        let longToken       = this.tokenManager.genLongToken({userId: createdUser._id, userKey: createdUser.key });
+        let longToken       = this.tokenManager.genLongToken({userId: createdUser._id });
         
         // Response
         return {
@@ -41,8 +46,13 @@ module.exports = class User {
         return { error: 'An error occurred while creating the user' };
     }
     }
-    async getUser({ userId }) {
+    async getUser({ __headers, userId }) {
         try {
+            // Make sure only user can fetch his own information
+            const decoded = this.tokenManager.verifyLongToken({ token: __headers.token });
+            if (!decoded || decoded.userId !== userId) {
+                return { error: 'Invalid or expired token' };
+            }
             let user = await this.mongomodels.UserModel.findById(userId);
             if (!user) {
                 return { error: 'User not found' };
@@ -52,8 +62,13 @@ module.exports = class User {
             return { error: 'An error occurred while retrieving the user' };
         }
     }
-    async updateUser({ userId, updates }) {
+    async updateUser({ __headers,userId, updates }) {
         try {
+            // Make sure only user can update his own information
+            const decoded = this.tokenManager.verifyLongToken({ token: __headers.token });
+            if (!decoded || decoded.userId !== userId) {
+                return { error: 'Invalid or expired token' };
+            }
             let user = await this.mongomodels.UserModel.findByIdAndUpdate(userId, updates, { new: true });
             if (!user) {
                 return { error: 'User not found' };
@@ -63,8 +78,13 @@ module.exports = class User {
             return { error: 'An error occurred while updating the user' };
         }
     }
-    async deleteUser({ userId }) {
+    async deleteUser({__headers, userId}) {
         try {
+            // Make sure only user can delete his own information
+            const decoded = this.tokenManager.verifyLongToken({ token: __headers.token });
+            if (!decoded || decoded.userId !== userId) {
+                return { error: 'Invalid or expired token' };
+            }
             let user = await this.mongomodels.UserModel.findByIdAndDelete(userId);
             if (!user) {
                 return { error: 'User not found' };
@@ -74,14 +94,67 @@ module.exports = class User {
             return { error: 'An error occurred while deleting the user' };
         }
     }
-    async listUsers({ page = 1, limit = 10 }) {
+    async listUsers({__headers, page = 1, limit = 10 }) {
         try {
+            const decoded = this.tokenManager.verifyLongToken({ token: __headers.token });
+            if (!decoded ) {
+                return { error: 'Invalid or expired token' };
+            }
+            let user = await this.mongomodels.UserModel.findById(decoded.userId);
+            //only superadmins can see the list of users
+            if(user.role !== 'superAdmin'){
+                return { error: 'You are not authorized to perform this operation' };
+            }
             let users = await this.mongomodels.UserModel.find()
                 .skip((page - 1) * limit)
                 .limit(limit);
             return { users };
         } catch (error) {
+            console.log(error);
             return { error: 'An error occurred while listing the users' };
+        }
+    }
+    async loginUser(req) {
+        try {
+            const { username, password } = req;
+            // Find the user by username
+            let user = await this.mongomodels.UserModel.findOne({ username });
+            if (!user) {
+                return { error: 'Invalid username or password' };
+            }
+    
+            // Validate the password
+            let isPasswordValid = await this.utils.comparePassword(password, user.password);
+            if (!isPasswordValid) {
+                return { error: 'Invalid username or password' };
+            }
+
+            if(superAdmins.includes(user.email)){
+                user.role = 'superAdmin'
+                await user.save(); 
+            }
+    
+            // Generate tokens
+            let longToken = this.tokenManager.genLongToken({ userId: user._id, userKey: user.key });
+            let shortTokenResponse = this.tokenManager.v1_createShortTokenFromLongToken({
+                __longToken: longToken,
+                __device: req.__device
+            });
+            
+            if (shortTokenResponse.error) {
+                return { error: 'An error occurred while generating short token' };
+            }
+            
+            let shortToken = shortTokenResponse.shortToken;
+    
+            // Respond with tokens
+            return {
+                user,
+                longToken,
+                shortToken
+            };
+        } catch (error) {
+            return { error: 'An error occurred while logging in' };
         }
     }
 }
